% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMixture.R
\name{calcMixture}
\alias{calcMixture}
\title{Wrapper to calculate sensitivity profile}
\usage{
calcMixture(sensitivity.tibb, drug.names = NULL,
  prior_proportion = 0.8, sd_na_val = 0.02)
}
\arguments{
\item{sd_na_val}{}
}
\value{
List of tibbles; one containing posterior estimates and one containing estimated parameters
}
\description{
A wrapper to return the sensitivity information for the
intersected PSets across cells and drugs.
}
\seealso{
\code{\link{intersectPSetWrapper}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{Sensitivity}
\keyword{intersectPSet}
