% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMLE.R
\name{calcMLE}
\alias{calcMLE}
\title{Wrapper to calculate sensitivity profile}
\usage{
calcMLE(sensitivity.tibb, drug.names = NULL, fit = "normal")
}
\arguments{
\item{drug.names}{List of drugs to fit model to}

\item{fit}{Return MLE for beta or normal distribution}

\item{sensitivitiy.tibb}{Tibble containing sensitivity information, drugs, and cells}
}
\value{
List of tibbles; one containing posterior estimates and one containing estimated parameters
}
\description{
A wrapper to return the sensitivity information for the
intersected PSets across cells and drugs.
}
\seealso{
\code{\link{intersectPSetWrapper}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{Sensitivity}
\keyword{intersectPSet}
