% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSensitivities.R
\name{computeSensitivities}
\alias{computeSensitivities}
\title{Wrapper to calculate and return a sensitivity profile}
\usage{
computeSensitivities(common, sensitivity_measure, truncateViabilities = TRUE)
}
\arguments{
\item{common}{List of \code{PSets}}

\item{sensitivity_measure}{Sensitivity measure to return}

\item{truncateViabilities}{Truncate \code{viabilities>100}}
}
\value{
Tibble containing sensitivity information
}
\description{
A wrapper to return the sensitivity information for the
intersected PSets across cells and drugs. The sensitivity measure must not be present
in the \code{PSet} but calculable from the dose-response information
(e.g. \code{'ec50', 'e_inf', 'hill_slope','raw_auc'}).
}
\seealso{
\code{\link{intersectPSetWrapper}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{Sensitivity}
\keyword{intersectPSet}
