% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMLE.R
\name{fit_bb_mle}
\alias{fit_bb_mle}
\title{Wrapper to calculate sensitivity profile}
\usage{
fit_bb_mle(x, n)
}
\arguments{
\item{x}{Number of sucessess in beta-binomial}

\item{n}{Number of trials}
}
\value{
List of parameter estimates
}
\description{
A wrapper to return the sensitivity information for the
intersected PSets across cells and drugs.
}
\seealso{
\code{\link{iterate_em}} \code{\link{fit_beta_mle}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{beta}
\keyword{binomial}
\keyword{mle}
