% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMLE.R
\name{fit_normal_mle}
\alias{fit_normal_mle}
\title{Wrapper to calculate sensitivity profile}
\usage{
fit_normal_mle(x)
}
\arguments{
\item{x}{Observed value}
}
\value{
List of parameter estimates
}
\description{
A wrapper to return the sensitivity information for the
intersected PSets across cells and drugs.
}
\seealso{
\code{\link{iterate_em}} \code{\link{fit_bb_mle}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{mle}
\keyword{normal}
