% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSentivities.R
\name{formatSensitivities}
\alias{formatSensitivities}
\title{Wrapper to return pre-computed sensitivity profile}
\usage{
formatSensitivities(common, sensitivity_measure)
}
\arguments{
\item{common}{List of \code{PSets}}

\item{sensitivity_measure}{Sensitivity measure to return}
}
\value{
Tibble containing sensitivity information
}
\description{
A wrapper to return the sensitivity information for the
intersected PSets across cells and drugs. The sensitivity measure must be present
in the \code{PSet} to utilize this function (e.g. \code{'auc_recomputed'}).
}
\seealso{
\code{\link{intersectPSetWrapper}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{Sensitivity}
\keyword{intersectPSet}
