% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectPSetWrapper.R
\name{intersectPSetWrapper}
\alias{intersectPSetWrapper}
\title{Wrapper to intersect PSets}
\usage{
intersectPSetWrapper(datasets, minimum_intersection = 1,
  intersectOn = c("cell.lines", "drugs"),
  sensitivity_measure = "auc_recomputed", remove_noisy_curves = FALSE,
  remove_mislabelled_cells = FALSE, mislabelled_cells = NULL)
}
\arguments{
\item{datasets}{Vector of datasets to calculate intersections among}

\item{minimum_intersection}{Number of experiments to calculate intersection across}

\item{intersectOn}{Vetor of items to perform intersection across. Default is \code{c'cell.lines','drugs')}}

\item{sensitivity_measure}{Vector of sensitivity measures to return}

\item{remove_noisy_curves}{Remove noisy dose-response curves}

\item{remove_mislabelled_cells}{Remove incorrectly-labelled cells}

\item{mislabelled_cells}{Vector of cells to remove. If \code{NULL}, GDSC and CCLE mislabelled cells used.}
}
\value{
List of tibbles containing sensitivity information
}
\description{
A wrapper to return the intersected PSets for cells and drugs
attaining a minimum number of intersections.
}
\seealso{
\code{\link{formatSensitivities}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{PSet}
\keyword{intersectPSet}
