% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_em.R
\name{iterate_em}
\alias{iterate_em}
\title{Wrapper to calculate beta mixture model}
\usage{
iterate_em(state, ...)
}
\arguments{
\item{state}{}
}
\value{
List of tibbles; one containing posterior estimates and one containing estimated parameters
}
\description{
A wrapper to return the sensitivity information for the
intersected PSets across cells and drugs.
}
\seealso{
\code{\link{intersectPSetWrapper}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{fit_bb_mle}
\keyword{fit_beta_mle}
