% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBetaMixture.R
\name{calcBetaMixture}
\alias{calcBetaMixture}
\title{Wrapper to calculate beta mixture model}
\usage{
calcBetaMixture(sensitivity.tibb, drug.names = NULL,
  prior_proportion = 0.8)
}
\arguments{
\item{prior_proportion}{}
}
\value{
List of tibbles; one containing posterior estimates and one containing estimated parameters
}
\description{
A wrapper to return the sensitivity information for the
intersected PSets across cells and drugs.
}
\seealso{
\code{\link{intersectPSetWrapper}}
}
\author{
Matthew Ploenzke, \email{ploenzke@g.harvard.edu}
}
\keyword{Sensitivity}
\keyword{intersectPSet}
